+++ controllers/reports.controller.js
@@ -1,7 +1,7 @@
 const asyncHandler = require('express-async-handler');
 const Report = require('../models/report.model');
 const Player = require('../models/player.model');
-const Order = require('../models/order.model');
+const Order = require('../modules/orders/order.model');
 const ApiError = require('../utils/ApiError');
 const prepareQuery = require('../utils/prepareQuery');
 const getIndividualSkillsProps = require('../utils/getIndividualSkillsProps');
