+++ scripts/seeder.js
@@ -8,7 +8,7 @@ dotenv.config({ path: './config/config.env' });
 const User = require('../models/user.model');
 const Club = require('../models/club.model');
 const Player = require('../models/player.model');
-const Order = require('../models/order.model');
+const Order = require('../modules/orders/order.model');
 const Report = require('../models/report.model');
 
 mongoose.connect(process.env.DB_CONNECT, {
