+++ test/db-utils.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcryptjs');
 const User = require('../models/user.model');
 const Club = require('../models/club.model');
 const Player = require('../models/player.model');
-const Order = require('../models/order.model');
+const Order = require('../modules/orders/order.model');
 const Report = require('../models/report.model');
 const { buildUser } = require('./utils');
